Backend/requirements.txt
flask==2.3.2
google-generativeai==0.6.0
python-dotenv==1.0.1
flask-cors==4.0.1